mensajes de error creados

+++ app/Http/Controllers/Estudiante/TorneoController.php
@@ -108,6 +108,8 @@ class TorneoController extends Controller
         //si no se ingresa ningun equipo que retorne a la pagina sin hacer nada
         //si el torneo esta lleno no puede ingresar
 
+
+
         if($request->equipoLiderado_id == null)
         {
             dd("no ha ingresado ningun equipo. Retorne a la página anterior.");
@@ -117,36 +119,39 @@ class TorneoController extends Controller
             $torneo = Torneo::find($request->torneo_id);
             $equipoLiderado = Equipo::find($request->equipoLiderado_id);
             $torneosEquipoLiderado = $equipoLiderado->torneos;
+            $consulta_1 = $torneosEquipoLiderado->where('modalidad_id', $equipoLiderado->modalidad_id);
+            $consulta_2 = $consulta_1->where('finalizado', 1);
 
             //si el torneo está cerrado no se puede entrar
             if($torneo->cerrado){
-                dd("el torneo está cerrado. Retorne a la página anterior.");
+                flash('el torneo está cerrado')->error();
             }
 
             //si ya pertenece al torneo no puede volver a ingresar
-            if(count($equipoLiderado->torneos->where('id', $torneo->id)) > 0)
+            elseif(count($equipoLiderado->torneos->where('id', $torneo->id)) > 0)
             {
-                dd("ya pertenece al torneo");
+                flash('Ya pertenece al torneo')->error();
             }
             //si está fuera de fecha no puede puede entrar
-            if(strtotime($torneo->fecha) < time())
+            elseif(strtotime($torneo->fecha) < time())
             {
-                dd("está fuera de plazo. Retorne a la página anterior.");
+                flash('está fuera de plazo')->error();
             }
 
             //si el torneo esta lleno no puede ingresar
-            if($torneo->max <= count($torneo->equipos))
+            elseif($torneo->max <= count($torneo->equipos))
             {
-                dd("el torneo está lleno. Retorne a la página anterior");
+                flash('el torneo está lleno')->error();
             }
             //si ya está inscrito en otro torneo de la misma modalidad no puede ingresar
-            $consulta_1 = $torneosEquipoLiderado->where('modalidad_id', $equipoLiderado->modalidad_id);
-            $consulta_2 = $consulta_1->where('finalizado', 1);
-            if(count($consulta_2)>0){
-                dd("ya pertenece a un torneo. Retorne a la página anterior.");
+            elseif(count($consulta_2)>0)
+            {
+                flash('ya pertenece a un torneo')->error();
             }
-
-            $torneo->equipos()->attach($equipoLiderado->id);
+            else{
+                flash('Se ha inscrito al torneo con éxito!')->success();
+                $torneo->equipos()->attach($equipoLiderado->id);
+            }
 
             return Redirect(route('estudiante.torneos.show', ['id' => $torneo->id]));
         }
@@ -157,14 +162,20 @@ class TorneoController extends Controller
     {
         if($request->equipoBaja_id == null)
         {
-            dd("no existe o no has seleccionado ningun equipo. Retorne a la página anterior");
+            flash('No existe o no has seleccionado ningun equipo')->error();
         } else
         {
             $torneo = Torneo::find($request->torneo_id);
-            $torneo->equipos()->detach($request->equipoBaja_id);
 
-            return Redirect(route('estudiante.torneos.show', ['id' => $torneo->id]));
+            //si el torneo está cerrado no puede salir
+            if($torneo->cerrado == 1){
+                flash('El torneo está cerrado, no puedes salirte ahora')->error();
+            } else{
+                $torneo->equipos()->detach($request->equipoBaja_id);
+                flash('Se ha dado de baja con éxito')->success();
+            }
         }
+        return Redirect(route('estudiante.torneos.show', ['id' => $request->torneo_id]));
     }
     /**
      * Show the form for editing the specified resource.
